+++ nengo/cache.py
@@ -39,9 +39,12 @@ from nengo.neurons import (
     Izhikevich,
     LIF,
     LIFRate,
+    PoissonSpiking,
     RectifiedLinear,
+    RegularSpiking,
     Sigmoid,
     SpikingRectifiedLinear,
+    StochasticSpiking,
     Tanh,
 )
 from nengo.rc import rc
@@ -197,9 +200,12 @@ class Fingerprint:
         Izhikevich,
         LIF,
         LIFRate,
+        PoissonSpiking,
         RectifiedLinear,
+        RegularSpiking,
         Sigmoid,
         SpikingRectifiedLinear,
+        StochasticSpiking,
         Tanh,
     )
     DISTRIBUTIONS = (