+++ nengo/tests/test_cache.py
@@ -61,6 +61,7 @@ neuron_types = [
     obj
     for obj in list_objects(nengo.neurons)
     if isstrictsubclass(obj, nengo.neurons.NeuronType)
+    and obj is not nengo.neurons.RatesToSpikesNeuronType
 ]
 solver_types = [
     obj
@@ -368,7 +369,12 @@ def test_supported_fingerprinting(cls, monkeypatch):
     monkeypatch.setitem(sys.modules, "sklearn.utils", Mock())
     monkeypatch.setitem(sys.modules, "sklearn.utils.extmath", Mock())
 
-    obj = cls()
+    args = []
+    if issubclass(cls, nengo.neurons.RatesToSpikesNeuronType):
+        # spiking types require a `base_type` argument, so provide one
+        args.append(nengo.neurons.RectifiedLinear())
+
+    obj = cls(*args)
     assert Fingerprint.supports(obj)
 
     # check fingerprint is created without error and is a valid sha1 hash